// frontend/src/lib/types/item-types.gen.ts
export const ItemTypes = {
  Normal: 0,
  Equip: 1,
  Skill: 2,
  Tent: 3,
  Installation: 4,
  Jewel: 5,
  CannonBall: 6,
  Mapae: 7,
  Material: 8,
  Interaction: 9,
  ContentsEvent: 10,
  ToVehicle: 11,
  Reserved12: 12,
  Reserved13: 13,
  Reserved14: 14,
  Reserved15: 15,
  Reserved16: 16,
  Reserved17: 17,
  Reserved18: 18,
  Reserved19: 19,
} as const;

export type ItemType = (typeof ItemTypes)[keyof typeof ItemTypes];

export interface ItemTypeInfo {
  value: ItemType;
  name: string;
  title: string;
  reserved?: boolean;
}

export const ItemTypeInfos: Record<ItemType, ItemTypeInfo> = {
  [ItemTypes.Normal]: {
    value: ItemTypes.Normal,
    name: "Normal",
    title: "General",
  },
  [ItemTypes.Equip]: {
    value: ItemTypes.Equip,
    name: "Equip",
    title: "Equipment",
  },
  [ItemTypes.Skill]: {
    value: ItemTypes.Skill,
    name: "Skill",
    title: "Consumable",
  },
  [ItemTypes.Tent]: {
    value: ItemTypes.Tent,
    name: "Tent",
    title: "Holding Tool",
  },
  [ItemTypes.Installation]: {
    value: ItemTypes.Installation,
    name: "Installation",
    title: "Installable Object",
  },
  [ItemTypes.Jewel]: {
    value: ItemTypes.Jewel,
    name: "Jewel",
    title: "Socket Item",
  },
  [ItemTypes.CannonBall]: {
    value: ItemTypes.CannonBall,
    name: "CannonBall",
    title: "Cannonball",
  },
  [ItemTypes.Mapae]: {
    value: ItemTypes.Mapae,
    name: "Mapae",
    title: "License",
  },
  [ItemTypes.Material]: {
    value: ItemTypes.Material,
    name: "Material",
    title: "Crafting Material",
  },
  [ItemTypes.Interaction]: {
    value: ItemTypes.Interaction,
    name: "Interaction",
    title: "Enter Area",
  },
  [ItemTypes.ContentsEvent]: {
    value: ItemTypes.ContentsEvent,
    name: "ContentsEvent",
    title: "Special Items",
  },
  [ItemTypes.ToVehicle]: {
    value: ItemTypes.ToVehicle,
    name: "ToVehicle",
    title: "General",
  },
  [ItemTypes.Reserved12]: {
    value: ItemTypes.Reserved12,
    name: "Reserved12",
    title: "General",
    reserved: true,
  },
  [ItemTypes.Reserved13]: {
    value: ItemTypes.Reserved13,
    name: "Reserved13",
    title: "General",
    reserved: true,
  },
  [ItemTypes.Reserved14]: {
    value: ItemTypes.Reserved14,
    name: "Reserved14",
    title: "General",
    reserved: true,
  },
  [ItemTypes.Reserved15]: {
    value: ItemTypes.Reserved15,
    name: "Reserved15",
    title: "General",
    reserved: true,
  },
  [ItemTypes.Reserved16]: {
    value: ItemTypes.Reserved16,
    name: "Reserved16",
    title: "General",
    reserved: true,
  },
  [ItemTypes.Reserved17]: {
    value: ItemTypes.Reserved17,
    name: "Reserved17",
    title: "General",
    reserved: true,
  },
  [ItemTypes.Reserved18]: {
    value: ItemTypes.Reserved18,
    name: "Reserved18",
    title: "General",
    reserved: true,
  },
  [ItemTypes.Reserved19]: {
    value: ItemTypes.Reserved19,
    name: "Reserved19",
    title: "General",
    reserved: true,
  },
} as const;